move config pkg under internal/;matchmaker_config.yaml is a part of Helm chart config now
