--- config/matchmaker_config.yaml
@@ -1,90 +1,0 @@
-# kubectl create configmap om-configmap --from-file=config/matchmaker_config.yaml
-debug: true
-
-logging:
-  level: debug
-  format: text
-  source: false
-
-api:
-  backend:
-    hostname: om-backendapi
-    port: 50505
-    backoff: "[2 32] *2 ~0.33 <30"
-    proxyport: 51505
-  frontend:
-    hostname: om-frontendapi
-    port: 50504
-    backoff: "[2 32] *2 ~0.33 <300"
-    proxyport: 51504
-  mmlogic:
-    hostname: om-mmlogicapi
-    port: 50503
-    proxyport: 51503
-  functions:
-    port: 50502
-    proxyport: 51502
-
-evaluator:
-  # Evaluator intervals are in milliseconds
-  pollIntervalMs: 1000
-  maxWaitMs: 10000
-
-metrics:
-  port: 9555
-  endpoint: /metrics
-  reportingPeriod: 5
-
-queues:
-  proposals:
-    name: proposalq
-
-ignoreLists:
-  proposed:
-    name: proposed
-    offset: 0
-    duration: 800
-  deindexed:
-    name: deindexed
-    offset: 0
-    duration: 800
-  expired:
-    name: OM_METADATA.accessed
-    offset: 800
-    duration: 0
-
-redis:
-  pool:
-    maxIdle: 3
-    maxActive: 0
-    idleTimeout: 60
-  queryArgs:
-    count: 10000
-  results:
-    pageSize: 10000
-  expirations:
-    player: 43200
-    matchobject: 43200
-
-jsonkeys:
-  rosters: properties.rosters
-  pools: properties.pools
-
-playerIndices:
-- char.cleric
-- char.knight
-- char.paladin
-- map.aleroth
-- map.oasis
-- mmr.rating
-- mode.battleroyale
-- mode.ctf
-- mode.demo
-- region.europe-east1
-- region.europe-west1
-- region.europe-west2
-- region.europe-west3
-- region.europe-west4
-- role.dps
-- role.support
-- role.tank
