--- install/helm/open-match/config/matchmaker_config.yaml
@@ -1,1 +1,0 @@
-../../../../config/matchmaker_config.yaml
