+++ internal/config/config_test.go
@@ -18,10 +18,19 @@ limitations under the License.
 package config
 
 import (
+	"io/ioutil"
+	"os"
 	"testing"
+	"time"
 )
 
 func TestReadConfig(t *testing.T) {
+	yaml := []byte(`metrics.endpoint: /metrics`)
+	if err := ioutil.WriteFile("matchmaker_config.yaml", yaml, 0666); err != nil {
+		t.Fatalf("could not create config file: %s", err)
+	}
+	defer os.Remove("matchmaker_config.yaml")
+
 	cfg, err := Read()
 	if err != nil {
 		t.Fatalf("cannot load config, %s", err)
@@ -29,5 +38,16 @@ func TestReadConfig(t *testing.T) {
 
 	if cfg.GetString("metrics.endpoint") != "/metrics" {
 		t.Errorf("av.GetString('metrics.endpoint') = %s, expected '/metrics'", cfg.GetString("metrics.endpoint"))
+	}
+
+	yaml = []byte(`metrics.endpoint: ''`)
+	if err := ioutil.WriteFile("matchmaker_config.yaml", yaml, 0666); err != nil {
+		t.Fatalf("could not update config file: %s", err)
+	}
+
+	time.Sleep(2 * time.Second)
+
+	if cfg.GetString("metrics.endpoint") != "" {
+		t.Errorf("av.GetString('metrics.endpoint') = %s, expected ''", cfg.GetString("metrics.endpoint"))
 	}
 }
