+++ internal/harness/matchfunction/golang/apisrv/apisrv.go
@@ -26,7 +26,7 @@ import (
 	"io"
 	"time"
 
-	"github.com/GoogleCloudPlatform/open-match/config"
+	"github.com/GoogleCloudPlatform/open-match/internal/config"
 	"github.com/GoogleCloudPlatform/open-match/internal/pb"
 	"github.com/gogo/protobuf/proto"
 	"go.opencensus.io/stats"
