+++ internal/serving/omserver.go
@@ -1,7 +1,7 @@
 package serving
 
 import (
-	"github.com/GoogleCloudPlatform/open-match/config"
+	"github.com/GoogleCloudPlatform/open-match/internal/config"
 	"github.com/GoogleCloudPlatform/open-match/internal/signal"
 	"github.com/gomodule/redigo/redis"
 	log "github.com/sirupsen/logrus"
