+++ internal/statestorage/redis/ignorelist/ignorelist.go
@@ -25,7 +25,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/GoogleCloudPlatform/open-match/config"
+	"github.com/GoogleCloudPlatform/open-match/internal/config"
 	"github.com/gomodule/redigo/redis"
 	log "github.com/sirupsen/logrus"
 )
